requirements.txt
amqp==5.1.1
asgiref==3.6.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.7
cron-descriptor==1.3.0
Django==4.2.1
django-bootstrap-v5==1.0.11
django-celery-beat==2.5.0
django-timezone-field==5.0
exceptiongroup==1.1.1
freezegun==1.2.2
idna==3.4
iniconfig==2.0.0
kombu==5.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
responses==0.23.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.9
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
